Week 5 - example projects
LetUsDance shows how to create a custom function and use it in various Sprites.  Also, how to create a complex dance project with multiple sprites is shown.  ChristmasTreeDecorater is an interesting project that is simple but uses elaborate and numerous costumes to provide a drag and drop Christmas tree decoration