Lesson 5 Coding Example - Learning Functions
This is a simple program that demonstrates the use of functions in python to carry out a repetitive operation.  In this example, the Happy Birthday song is reused multiple times with various students' names passed to the function.  Without the function, one will have to type 4 print statements for each student name!