Lesson 5 - Beginning Class Project
This program uses the concept of functions learned this week and also shows students how to scramble a sentence into a secret message that no one will be able to decipher.  This program only shows the first part of encryption.  In the class we also played with a different program that shows the decryption part.  The next few weeks will focus on teaching the concepts so that students will learn to encrypt and decrypt their own messages.  The cipher used is very simple (Caesar Cipher) where the alphabets are moved "n" places forward.  How many places a student wants to move the alphabet is determined by a key the student picks.