Lesson 6 - coding example of a dictionary
This example shows how to use a dictionary in python.  Students learned that a dictionary has keys and values.  Each key is mapped to a value.  In our example, we mapped an animal name to the sound it makes.  When running the program, a student can provide the name of an animal from the available list of animals and the program will retrieve the corresponding animal sound from the dictionary.